backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
xgboost==2.1.1
numpy==2.0.2
python-dotenv==1.0.1
